refator: move classic map data
