--- src/maps/raw/territory/README.md
@@ -1,1 +1,0 @@
-looked up manually on the internet (wiki + census 2010 data)
