+++ src/modes/classic/maps/index.js
@@ -1,5 +1,4 @@
-export * from './infos';
-import { nationInfo, stateInfo, countyInfo, msaInfo, hrrInfo, levelMegaCountyId, hhsInfo } from './infos';
+import { nationInfo, stateInfo, countyInfo, msaInfo, hrrInfo, levelMegaCountyId, hhsInfo } from '../../../maps/infos';
 
 export function loadSources(additionalProperties = {}) {
   // mark to be loaded as fast as possible
