+++ src/modes/classic/maps/raw/adjacency/README.md
