+++ src/modes/classic/maps/raw/adjacency/county_adjacency.txt
