+++ src/modes/classic/maps/raw/name_id_info.json
