+++ src/modes/classic/maps/raw/territory/README.md
@@ -1,0 +1,1 @@
+looked up manually on the internet (wiki + census 2010 data)
