--- src/specs/tilegrams/.gitignore
@@ -1,1 +1,0 @@
-/tiles.*
