--- src/specs/tilegrams/README.md
@@ -1,13 +1,0 @@
-generated using https://pitchinteractiveinc.github.io/tilegrams/
-
-## US States
-
- 1. Go to https://pitchinteractiveinc.github.io/tilegrams/
- 1. select US
- 1. Select NPR 1-to-1
- 1. Download topojson and store as tiles.topo.json in this directory
- 1. Go back to the website and drag/place *Florida* tile to where Puerto Rico should be
- 1. Download topojson and store as tiles.topo.shifted.json in this directory
- 1. run `node process.js` to convert to `state.topo.json`
-
-
