--- src/specs/tilegrams/nation.topo.json
@@ -1,1 +1,0 @@
-{"type":"Topology","objects":{"nation":{"type":"GeometryCollection","geometries":[{"type":"MultiPolygon","arcs":[[[0]],[[1]],[[2]],[[3]]],"id":"us"}]}},"arcs":[[[292.2033862398606,33.74074074242778],[321.4237248667687,16.870370370370374],[350.6440634936768,33.74074074242778],[350.6440634936768,67.4814814865426],[321.4237248667687,84.35185185860001],[292.2033862398606,67.4814814865426],[292.2033862398606,33.74074074242778]],[[379.8644021205849,185.57407409094446],[409.08474074749296,168.70370371888706],[409.08474074749296,134.96296297477224],[438.3050793744011,118.09259260271483],[467.52541800130916,134.96296297477224],[496.74575662821724,118.09259260271483],[496.74575662821724,84.35185185860001],[525.9660952551253,67.4814814865426],[555.1864338820335,84.35185185860001],[584.4067725089415,67.4814814865426],[584.4067725089415,33.74074074242778],[613.6271111358496,16.870370370370374],[642.8474497627577,33.74074074242778],[642.8474497627577,67.4814814865426],[672.067788316615,84.35185185860001],[701.288126943523,67.4814814865426],[730.5084655704311,84.35185185860001],[759.7288041973393,67.4814814865426],[759.7288041973393,33.74074074242778],[788.9491428242472,16.870370370370374],[818.1694814511554,33.74074074242778],[818.1694814511554,67.4814814865426],[847.3898200780635,84.35185185860001],[847.3898200780635,118.09259260271483],[876.6101587049716,134.96296297477224],[905.8304973318797,118.09259260271483],[935.0508359587877,134.96296297477224],[935.0508359587877,168.70370371888706],[964.2711745856958,185.57407409094446],[964.2711745856958,219.31481483505928],[993.491513212604,236.18518516494078],[993.491513212604,269.9259259090556],[1022.7118518395121,286.796296281113],[1022.7118518395121,320.5370370252278],[993.491513212604,337.4074073972852],[993.491513212604,371.14814814140004],[1022.7118518395121,388.0185185134574],[1022.7118518395121,421.75925925757224],[993.491513212604,438.6296296296297],[964.2711745856958,421.75925925757224],[964.2711745856958,388.0185185134574],[935.0508359587877,371.14814814140004],[905.8304973318797,388.0185185134574],[876.6101587049716,371.14814814140004],[876.6101587049716,337.4074073972852],[847.3898200780635,320.5370370252278],[847.3898200780635,286.796296281113],[818.1694814511554,269.9259259090556],[788.9491428242472,286.796296281113],[788.9491428242472,320.5370370252278],[759.7288041973393,337.4074073972852],[730.5084655704311,320.5370370252278],[730.5084655704311,286.796296281113],[701.288126943523,269.9259259090556],[672.067788316615,286.796296281113],[672.067788316615,320.5370370252278],[642.8474497627577,337.4074073972852],[613.6271111358496,320.5370370252278],[584.4067725089415,337.4074073972852],[555.1864338820335,320.5370370252278],[525.9660952551253,337.4074073972852],[496.74575662821724,320.5370370252278],[467.52541800130916,337.4074073972852],[438.3050793744011,320.5370370252278],[409.08474074749296,337.4074073972852],[379.8644021205849,320.5370370252278],[379.8644021205849,286.796296281113],[409.08474074749296,269.9259259090556],[409.08474074749296,236.18518516494078],[379.8644021205849,219.31481483505928],[379.8644021205849,185.57407409094446]],[[321.4237248667687,388.0185185134574],[350.6440634936768,371.14814814140004],[379.8644021205849,388.0185185134574],[379.8644021205849,421.75925925757224],[350.6440634936768,438.6296296296297],[321.4237248667687,421.75925925757224],[321.4237248667687,388.0185185134574]],[[935.0508359587877,33.74074074242778],[964.2711745856958,16.870370370370374],[993.491513212604,33.74074074242778],[993.491513212604,67.4814814865426],[964.2711745856958,84.35185185860001],[935.0508359587877,67.4814814865426],[935.0508359587877,33.74074074242778]]],"bbox":[292.2033862398606,16.870370370370374,1022.7118518395121,438.6296296296297]}
